+ README updated + Light class (Point Light) created, holding the position and color of the light + Material parent class added to project, has two children, a normal monotone color, and a checkered variant, their properties dictate their glossiness and etc. = Engine : Specular and Diffuse lighting, recursively checks ray reflections up to a previously declared depth - TODO : primitive shadow calculation in lines 103 and 104 of Engine.cpp is faulty. must check before progressing. = 3DObjects : Addec a function to get their normal vector. = Ray : Added an option to store both normalized and untouched vectors for the direction. = Scene : changed to store lights. - Renamed Sphere to Object3DSphere
NOTE : The continuation of development is halted to address some serious bugs regarding directions of vectors, camera position and its direction, light ray directions that result in inconsistent and buggy object reflection and light behaviours and etc. .Might have messed up some reference calls and addresses of a couple of objects when calling them in different classes. Rendering works for now but is unsatisfactory and faulty. Thus, other features won't be implemented untill these are fixed.